Refactor payment objects
